Code Refactoring: Class Structure and Interface Creation
Class Restructuring: In this commit, significant changes were made to the code structure. Interfaces "Board" and "Player" were created, alongside the abstract class "TypeOfGame," which serves as the parent class for the following three concrete classes: "HorizontalAndVerticalGame," "DiagonalGame," and "CompleteGame."

Temporary Draft: It's essential to note that these changes were made as a temporary draft. There is still work to be done since there are unnecessary conditional statements (ifs) that need to be removed, and class cleanliness needs improvement.

Code Cleanup: Furthermore, within the "Linea" class, both loops (for) and conditional statements (if) have been identified for removal in future commits to enhance code quality.